# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/users": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get all users",
                "operationId": "FindAll",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.Response"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.Response": {
            "type": "object",
            "properties": {
                "us_apellido": {
                    "type": "string"
                },
                "us_correo": {
                    "type": "string"
                },
                "us_eliminado": {
                    "type": "boolean"
                },
                "us_esactivo": {
                    "type": "boolean"
                },
                "us_fecharegistro": {
                    "type": "string"
                },
                "us_id": {
                    "type": "integer"
                },
                "us_nombre": {
                    "type": "string"
                },
                "us_usuario": {
                    "type": "string"
                }
            }
        }
    }
}